make this library a bagillion times faster on rebuilds.

stress test app: 2,000ms -> 130ms
one of my apps at work: 1,600ms -> 94ms